# pocket/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cotejo citation check",
    "version": "0.1.0",
    "description": "Citation integrity for AI agents. Checks whether a cited page is alive, whether a quote is on it, and, when it is not, whether the Wayback Machine copy closest to the citation date contained it. No LLM: matching is token-normalised text comparison. Served on Pocket Network as service `citation-check` (REST). Every response is a JSON object; all inputs arrive in the POST body.",
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "https://cotejo-demo.up.railway.app", "description": "Public demo, rate limited (no payment, for trying the API)" }
  ],
  "paths": {
    "/v1/check": {
      "post": {
        "operationId": "checkCitations",
        "summary": "Check up to 10 citations",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CheckRequest" },
            "example": { "items": [ { "url": "https://example.com", "quote": "This domain is for use in illustrative examples in documents", "cited_at": "2023-01-01" } ] } } }
        },
        "responses": {
          "200": { "description": "One result per item, in request order", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CheckResponse" } } } },
          "400": { "$ref": "#/components/responses/Error" },
          "413": { "$ref": "#/components/responses/Error" },
          "429": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/match": {
      "post": {
        "operationId": "matchQuote",
        "summary": "Match a quote against text you already have (no fetching, deterministic)",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["text", "quote"],
            "properties": { "text": { "type": "string", "maxLength": 200000 }, "quote": { "type": "string", "maxLength": 2000 } } },
            "example": { "text": "Pocket Network relays paid API calls.", "quote": "relays paid API calls" } } }
        },
        "responses": {
          "200": { "description": "Match result", "content": { "application/json": { "schema": { "type": "object", "required": ["service", "match"],
            "properties": { "service": { "type": "string", "enum": ["citation-check"] }, "match": { "$ref": "#/components/schemas/Match" } } } } } },
          "400": { "$ref": "#/components/responses/Error" },
          "413": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/verdicts": {
      "get": { "operationId": "listVerdicts", "summary": "Meaning of every verdict and archive_status value",
        "responses": { "200": { "description": "Verdict glossary", "content": { "application/json": { "schema": { "type": "object",
          "properties": { "verdicts": { "type": "object", "additionalProperties": { "type": "string" } }, "archive_status": { "type": "object", "additionalProperties": { "type": "string" } } } } } } } } }
    },
    "/v1/version": {
      "get": { "operationId": "version", "summary": "Identity probe",
        "responses": { "200": { "description": "Service id and version", "content": { "application/json": { "schema": { "type": "object",
          "properties": { "service": { "type": "string", "enum": ["citation-check"] }, "version": { "type": "string" } } } } } } } }
    },
    "/v1/health": {
      "get": { "operationId": "health", "summary": "Readiness probe",
        "responses": { "200": { "description": "Ready", "content": { "application/json": { "schema": { "type": "object",
          "properties": { "status": { "type": "string", "enum": ["ok"] } } } } } } } }
    }
  },
  "components": {
    "responses": {
      "Error": { "description": "Bad input or limit reached; never retried by gateways", "content": { "application/json": { "schema": { "type": "object", "required": ["error"],
        "properties": { "error": { "type": "object", "required": ["code"], "properties": { "code": { "type": "string" }, "detail": { "type": "string", "nullable": true } } } } } } } }
    },
    "schemas": {
      "CheckRequest": {
        "type": "object", "required": ["items"],
        "properties": {
          "items": { "type": "array", "minItems": 1, "maxItems": 10, "items": { "type": "object", "required": ["url"],
            "properties": {
              "url": { "type": "string", "maxLength": 2048, "description": "http or https, ports 80/443, public addresses only" },
              "quote": { "type": "string", "maxLength": 2000, "description": "Omit to check liveness only" },
              "cited_at": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$", "description": "Archive copy closest to this date is compared; without it, the latest copy" } } } },
          "archive": { "type": "string", "enum": ["auto", "always", "never"], "default": "auto" }
        }
      },
      "CheckResponse": {
        "type": "object", "required": ["service", "version", "checked_at", "summary", "results"],
        "properties": {
          "service": { "type": "string", "enum": ["citation-check"] },
          "version": { "type": "string" },
          "checked_at": { "type": "string", "format": "date-time" },
          "summary": { "type": "object", "required": ["total"], "properties": { "total": { "type": "integer" } }, "additionalProperties": { "type": "integer" } },
          "results": { "type": "array", "items": { "$ref": "#/components/schemas/Result" } }
        }
      },
      "Result": {
        "type": "object", "required": ["url", "verdict", "archive_status"],
        "properties": {
          "url": { "type": "string" },
          "quote": { "type": "string", "nullable": true },
          "verdict": { "type": "string", "enum": ["supported", "weak", "not_found", "drifted", "dead_archived", "dead_unverified", "dead", "alive", "unchecked", "rejected"] },
          "reason": { "type": "string", "description": "Present when verdict is rejected" },
          "archive_status": { "type": "string", "enum": ["checked", "none", "unavailable", "skipped"] },
          "live": { "$ref": "#/components/schemas/Fetch" },
          "archive": { "type": "object", "nullable": true, "properties": {
            "available": { "type": "boolean", "nullable": true },
            "timestamp": { "type": "string", "description": "Wayback timestamp YYYYMMDDhhmmss" },
            "snapshot_url": { "type": "string" },
            "http_status": { "type": "integer", "nullable": true },
            "sha256": { "type": "string", "nullable": true },
            "match": { "$ref": "#/components/schemas/Match" },
            "reason": { "type": "string" } } }
        }
      },
      "Fetch": {
        "type": "object", "nullable": true,
        "properties": {
          "reachable": { "type": "boolean" },
          "http_status": { "type": "integer" },
          "reason": { "type": "string", "description": "Why the page was not reachable, e.g. http_404, dns_not_found, no_response_in_time" },
          "final_url": { "type": "string", "nullable": true },
          "content_type": { "type": "string", "nullable": true },
          "bytes": { "type": "integer" },
          "truncated": { "type": "boolean" },
          "sha256": { "type": "string", "description": "SHA-256 of the fetched bytes" },
          "fetched_at": { "type": "string", "format": "date-time" },
          "match": { "$ref": "#/components/schemas/Match" }
        }
      },
      "Match": {
        "type": "object", "nullable": true,
        "properties": {
          "kind": { "type": "string", "enum": ["exact", "near", "partial", "absent", "unreadable"] },
          "score": { "type": "number", "nullable": true, "minimum": 0, "maximum": 1 },
          "excerpt": { "type": "string", "nullable": true },
          "quote_tokens": { "type": "integer" }
        }
      }
    }
  }
}